Land the current version of Snake in this new repository.
Snake is a popular game where the player has control of a snake on a 2d board. The objective is to eat food and grow the snake longer.
The challenge of this game comes in managing the ever growing trail of the snakes body.

The objective of this script is to allow a user to just play snake themselves, or using CLI flags have the game play itself to varying degrees of success.
The best algo currently is the A* algo, which does a good job of solving snake but will inevitably trap itself and fail.

Further work is needed to make the A* algo not trap the snake or allow the snake to get out if it is trapped.

new file:   snake.py